perf: Add python3-setuptools-native to perf dependencies to fix do_compile failure
Issue: LINCD-10571

This is actually for linux-yocto-dev which is v6.0 now. linux-yocto which is
v5.15 does not have this issue. The patch can be dropped once it is fixed
upstream. The patch is still under test but works fine for us for the moment.

(LOCAL REV: NOT UPSTREAM) -- Taken from oe-core mailing list
https://lists.openembedded.org/g/openembedded-core/message/171590

Signed-off-by: He Zhe <zhe.he@windriver.com>