qemu: Install the default qemu emulation rpm
Issue: LINUXEXEC-27688 LIN1023-907 LIN1023-912

The qemu rpm can be split or not via customize PACKAGESPLITFUNCS and
there is no specific qemu emulation rpm installed when we choose split
the qemu rpms now.

To gurantee the basic usage, install the qemu emulation rpm which
corresponding to the target arch by default when split the qemu rpm.

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core 20230615

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>