python3-requests: fix CVE-2023-32681
Issue:LIN1023-669

CVE: CVE-2023-32681

Requests is a HTTP library. Since Requests 2.3.0, Requests has been leaking
Proxy-Authorization headers to destination servers when redirected to an HTTPS
endpoint. This is a product of how we use `rebuild_proxies` to reattach the
`Proxy-Authorization` header to requests. For HTTP connections sent through
the tunnel, the proxy will identify the header in the request itself and remove
it prior to forwarding to the destination server. However when sent over HTTPS,
the `Proxy-Authorization` header must be sent in the CONNECT request as the
proxy has no visibility into the tunneled request. This results in Requests
forwarding proxy credentials to the destination server unintentionally, allowing
a malicious actor to potentially exfiltrate sensitive information. This issue
has been patched in version 2.31.0.

References:
https://github.com/psf/requests/security/advisories/GHSA-j8r2-6x86-q33q
https://github.com/psf/requests/commit/74ea7cf7a6a27a4eeb2ae24e162bcc942a6706d5

(mickledore rev: 11b6e64c07df043441824511c931fb0bc9673adc) -- Backport

Signed-off-by: Chee Yang Lee <chee.yang.lee@intel.com>
Signed-off-by: Steve Sakoman <steve@sakoman.com>
Signed-off-by: Narpat Mali <narpat.mali@windriver.com>