inetutils: fix CVE-2023-40303
Issue: LIN1023-1691

GNU inetutils through 2.4 may allow privilege escalation because
of unchecked return values of set*id() family functions in ftpd,
rcp, rlogin, rsh, rshd, and uucpd. This is, for example, relevant
if the setuid system call fails when a process is trying to drop
privileges before letting an ordinary user control the activities
of the process.

Refernces:
https://nvd.nist.gov/vuln/detail/CVE-2023-40303

CVE: CVE-2023-40303

(LOCAL REV: NOT UPSTREAM) -- sent to oe-core(mickledore) on 20230829

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>