tiff: fix CVE-2023-6228 and CVE-2023-6277
Issue: LIN1023-2758 LIN1023-2768

CVE-2023-6228:
An issue was found in the tiffcp utility distributed by the
libtiff package where a crafted TIFF file on processing may
cause a heap-based buffer overflow leads to an application
crash.

CVE-2023-6277:
An out-of-memory flaw was found in libtiff. Passing a crafted
tiff file to TIFFOpen() API may allow a remote attacker to
cause a denial of service via a craft input with size smaller
than 379 KB.

References:
https://nvd.nist.gov/vuln/detail/CVE-2023-6228
https://gitlab.com/libtiff/libtiff/-/issues/606
https://nvd.nist.gov/vuln/detail/CVE-2023-6277
https://gitlab.com/libtiff/libtiff/-/issues/614
https://gitlab.com/libtiff/libtiff/-/merge_requests/545

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>