ofono: fix CVE-2023-4234
Issue: LIN1023-4835

A flaw was found in ofono, an Open Source Telephony on Linux. A stack overflow bug
is triggered within the decode_submit_report() function during the SMS decoding.
It is assumed that the attack scenario is accessible from a compromised modem,
a malicious base station, or just SMS. There is a bound check for this memcpy
length in decode_submit(), but it was forgotten in decode_submit_report().

References:
https://nvd.nist.gov/vuln/detail/CVE-2023-4234

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>