ofono: fix CVE-2023-4233
A flaw was found in ofono, an Open Source Telephony on Linux. A stack overflow bug
is triggered within the sms_decode_address_field() function during the SMS
PDU decoding. It is assumed that the attack scenario is accessible from a compromised
modem, a malicious base station, or just SMS.

Issue: LIN1023-4834

References:
https://nvd.nist.gov/vuln/detail/CVE-2023-4233

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>