xserver-xorg: fix CVE-2024-31080
A heap-based buffer over-read vulnerability was found in the X.org server's
ProcXIGetSelectedEvents() function. This issue occurs when byte-swapped length
values are used in replies, potentially leading to memory leakage and segmentation
faults, particularly when triggered by a client with a different endianness.
This vulnerability could be exploited by an attacker to cause the X server to
read heap memory values and then transmit them back to the client until encountering
an unmapped page, resulting in a crash. Despite the attacker's inability to control
the specific memory copied into the replies, the small length values typically stored
in a 32-bit integer can result in significant attempted out-of-bounds reads.

Issue: LIN1023-4596

References:
https://nvd.nist.gov/vuln/detail/CVE-2024-31080

Upstream-patches:
https://gitlab.freedesktop.org/xorg/xserver/-/commit/96798fc1967491c80a4d0c8d9e0a80586cb2152b

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>