xserver-xorg: fix CVE-2024-31083
A use-after-free vulnerability was found in the ProcRenderAddGlyphs() function
of Xorg servers. This issue occurs when AllocateGlyph() is called to store new
glyphs sent by the client to the X server, potentially resulting in multiple
entries pointing to the same non-refcounted glyphs. Consequently, ProcRenderAddGlyphs()
may free a glyph, leading to a use-after-free scenario when the same glyph pointer is
subsequently accessed. This flaw allows an authenticated attacker to execute arbitrary
code on the system by sending a specially crafted request.

Issue: LIN1023-4599

References:
https://nvd.nist.gov/vuln/detail/CVE-2024-31083

Upstream-patches:
https://gitlab.freedesktop.org/xorg/xserver/-/commit/bdca6c3d1f5057eeb31609b1280fc93237b00c77
https://gitlab.freedesktop.org/xorg/xserver/-/commit/337d8d48b618d4fc0168a7b978be4c3447650b04

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Archana Polampalli <archana.polampalli@windriver.com>