curl: fix CVE-2024-2398
Issue: LIN1023-4428

When an application tells libcurl it wants to allow
HTTP/2 server push, and the amount of received headers
for the push surpasses the maximum allowed limit (1000),
libcurl aborts the server push. When aborting, libcurl
inadvertently does not free all the previously allocated
headers and instead leaks the memory. Further, this error
condition fails silently and is therefore not easily
detected by an application.

Reference:
https://nvd.nist.gov/vuln/detail/CVE-2024-2398

Upstream-patch:
https://github.com/curl/curl/commit/deca8039991886a559b67bcd6

(LOCAL REV: NOT UPSTREAM) -- Not applicable, upstream out of maintenance

Signed-off-by: Yogita Urade <yogita.urade@windriver.com>